WIP: Create initial LineChart component
